review: moved and renamed SupporterPlusLegal to helpers/legalCopy
