+++ support-frontend/assets/helpers/legalCopy.ts
